chore: extract user menu to separate component
